+++ app/javascript/users/mixins/homepage_content_mixin.js
@@ -3,6 +3,9 @@ import { mapState } from 'vuex'
 export const homepageContentMixin = {
   computed: {
     ...mapState('application', ['current_organization']),
+    isCurrentOrganizationPresent() {
+      return !!this.current_organization.name
+    },
     organizationHomepageData() {
       return this.current_organization.homepage
     },